fix(server): drop redundant uint32() conversion
Recovers work from commit e1f480e (lost in the author-fix rewrite).
golangci-lint's unconvert linter (visible after the v6 action
upgrade) flagged uint32(request.GetLimit()) as a redundant
conversion — GetLimit() already returns uint32 per the generated
proto code. Assign directly, no behavior change.

+++ internal/server/asset.go
@@ -484,7 +484,7 @@ func (s *Server) GetRecentlyAddedAssets(ctx context.Context, request *immichv1.G
 	userID := pgtype.UUID{Bytes: uid, Valid: true}
 
 	// Default to 12 when limit is 0, cap at 100 to prevent abuse
-	limit := uint32(request.GetLimit())
+	limit := request.GetLimit()
 	if limit == 0 {
 		limit = 12
 	}
